pyproject.toml
[project]
name = "telegram-faq-bot"
version = "0.1.0"
description = "Интеллектуальный Telegram-бот с AI для автоматизации общения с клиентами"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.3.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "beartype>=0.17.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8.0",
    "pyright>=1.1.350",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
